Fix links into grpc.io
As of today (2020/09/14), the second most popular 404 over [grpc.io](https://grpc.io) is the path `/grpc.github.io/docs/quickstart/python.html`.

I searched all of GitHub, and the page edited in this PR was the only one found to contain this invalid link to the Python quickstart. In fact, all links of the form `grpc.io/grpc.github.io` are invalid. Maybe these invalid links are the result of a search-and-replace error?

Anyhow, this PR fixes the links. I double checked with a link checker to ensure that the new links are all valid.